Deploy vsock/grpc backends
The idea of the 'aosp-xenvm-trout' device is to be executed on the Xen
hypervisor inside the DomA. This guest domain relies on existance of the
virtio devices, shared from the DomD domain.

But besides those devices the 'aosp-xenvm-trout' relies on the
availability of the several grpc/vsock backends. They are providing data
to the 'Audio Control HAL', 'Dumpstate HAL', 'Vehicle HAL'.

Those backends are packed by Google into the 'agl_services_build' Yocto
meta-layer. It should be built and deployed inside DomD.

The purpose of this patch it to copy and extend the agl_services_build
functionality from the 'trout' device in a way that it could be used
in DomD.

Keep in mind, that until vsock communication is not proved to work, we
leave local grpc server inside Android up and running. So it will be
running both in DomD and inside Android. That should do no harm as there
is only one client which will connect only to one of the available
sources.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>
Reviewed-by: Andrii Chepurnyi <andrii_chepurnyi@epam.com>