Override impl. of the virtualization-grpc-server
In order to have 'trout' running we need to have virtualization GRPC
server available. Vehicle hal relies on its presence. In the future, it
should be available in DomD. But as far as we are experimenting with
QEMU, we will deploy slightly extended ( gets more parameters from rc
file ) version of it right inside Android.

The following things are done:

- Added minimalistic override of the virtualization-grpc-server (
android.hardware.automotive.vehicle@2.0-virtualization-grpc-server-extended
) with modified list of parameters and changed service name
- Added related sepolicy rules
- Added service to the PRODUCT_PACKAGES

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>
Reviewed-by: Andrii Chepurnyi <andrii_chepurnyi@epam.com>