Added overriden impl. of the audio.primary.caremu
There is a crash in Android 12 version of the "audio.primary.caremu"
service implementation.

This patch introduces the overriden implementation of this service,
which fixes the crash.

The fix of the crash is located in the 'audio_hw.c' line 1221.
pthread_create call should be moved to a certain place in order to avoid
a race condition.

Initially the call was located here:

https://cs.android.com/android/platform/superproject/+/android-
12.1.0_r8:device/generic/car/emulator/audio/driver/audio_hw.c;l=1196

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>
Reviewed-by: Andrii Chepurnyi <andrii_chepurnyi@epam.com>