Remove cluster display port
We've started to observe that for some teammates the content of the
cluster display port was drawn on top of the main display port's
content. To step over this issue we've decided to turn off the cluster
display port.

This patch introduces an overlay, which overrides the car service's
'config_occupant_display_mapping' configuration setting.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>
Reviewed-by: Andrii Chepurnyi <andrii_chepurnyi@epam.com>