Rework display initialization
It has been discovered that a race condition sometimes
occurs between the HWC and the display module’s probe() function.
The only solution is to wait for the connector to appear
before starting the HWC. Since the HWC is started by Android,
it is necessary to wait some time for the connector to become available.
Note: The module probe can take a significant amount of time.
Additionally, use the HWC3 service implementation.

Signed-off-by: Andrii Chepurnyi <andrii_chepurnyi@epam.com>
Acked-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>